move storage into a dedicated folder
