+++ packages/extension/src/background/storage/interface.ts
